+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Manutenção: Limpeza da suite de testes do frontend e adicionada cobertura de testes no mesmo.
 * Funcionalidade: Adicionada validação de permissão ao criar e alterar reservas no servidor.
 * Correção: Corrigidos testes no servidor com mudança de Secret do Auth0.
+* Funcionalidade: Adicionado modal com informações sobre os desenvolvedores e links para github.
 
 # 0.0.2
 * Início versionamento.